fix install

+++ as-setup.sh
@@ -4,14 +4,15 @@ apt update
 apt install -y bc
 cd /tmp
 wget https://github.com/lexandr0s/autoswitch/raw/master/autoswitch.tar
-tar -xvf autoswitch.tar
+tar -xf autoswitch.tar 
 cp as/autoswitch /hive/sbin
 cp as/autoswitch_bin /hive/sbin
-cp as/autoswitch.conf /hive-config
-cp as/rig_data.json /hive-config
+[[ ! -f /hive-config/autoswitch.conf ]] && cp as/autoswitch.conf /hive-config
+[[ ! -f /hive-config/rig_data.json ]] && cp as/rig_data.json /hive-config
 rm -R as
 rm autoswitch.tar
 cd /home/user
+echo
 echo "Install Autoswitch complete"
 echo "Now you need to configure it."
 echo "See the manual"
